% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_well_data.R
\name{bind_well_data}
\alias{bind_well_data}
\title{bind_well_data: Binds well data using data files that are exported from
HoboWare as .csv files.}
\usage{
bind_well_data(path)
}
\arguments{
\item{path}{Quoted path of the folder where the exported Hobo tables are located.}
}
\value{
Returns a tdata frame with ID (blank for Access to assign autonumber), Well_ID,
Measure_Date_Time, Absolute_Pressure_kPa, Degrees_C, and site.
}
\description{
This function finds the well data files based on a specified path
and partial matching of files based on site name, reads these files into R,
adds Well_ID and site name to each file, then binds them together and prepares
the data to be imported into the MS Access RAM backend table: tbl_Water_Level.
If appending to the Access table is not needed, then drop the first column 'ID',
which is left blank so that Access can assign an auto number for the primary key.
}
\examples{
path='D:/NETN/Monitoring_Projects/Freshwater_Wetland/Hobo_Data/Fall_2018'
well_data<-bind_well_data(path)

}
