% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_WL_stats.R
\name{calc_WL_stats}
\alias{calc_WL_stats}
\title{calc_WL_stats: Calculates growing season water level statistics.}
\usage{
calc_WL_stats(df, from = 2013, to = 2019)
}
\arguments{
\item{df}{Data frame containing water level and precipitation data}

\item{from}{Year to start analysis, ranging from 2013-2019}

\item{to}{Year to stop analysis, ranging from 2013-2019}
}
\value{
Returns a data frame with water level growing season statistics.
}
\description{
This function calculates growing season water level statistics by site
and by year for sentinel wetlands in Acadia NP. Metrics follow Miller & Weed 2017
}
\examples{
well_data<-read.csv("well_prec_data_2013-2019.csv")
gs_wl_stats<-calc_WL_stats(df=well_data, from = 2017, to = 2019)

}
\seealso{
To view report that metrics are based on: \url{https://irma.nps.gov/DataStore/Reference/Profile/2239739}
}
