% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_WL_data.R
\name{compile_WL_data}
\alias{compile_WL_data}
\title{compile_WL_data: Converts well data to water level}
\usage{
compile_WL_data(
  path = NULL,
  year = as.numeric(format(Sys.Date(), "\%Y")),
  rejected = FALSE,
  growing_season = TRUE,
  export = TRUE,
  quietly = FALSE
)
}
\arguments{
\item{path}{Quoted path of the folder where the exported Hobo tables are located.}

\item{year}{Numeric. The year you are preparing the data for. Function will only run 1 year at a time.}

\item{rejected}{\code{TRUE} or \code{FALSE}. If \code{TRUE} will include flagged rejected records.
If \code{FALSE} will replace the rejected flagged records with NA. Defaults to \code{FALSE}.}

\item{growing_season}{\code{TRUE} or \code{FALSE}. If \code{TRUE} will include only growing season
measurements, defined as May 15 to Oct. 1. If \code{FALSE}, will return all measurements.
If growing_season = \code{TRUE} and export = \code{TRUE}, the output filename will include "GS"
Defaults to \code{TRUE}.}

\item{export}{\code{TRUE} or \code{FALSE}. Export csv file to specified path. Defaults to \code{TRUE}.}

\item{quietly}{\code{TRUE} or \code{FALSE}. If \code{FALSE}, code will not print progress into console.
Defaults to \code{FALSE}}
}
\value{
Returns a wide data frame with timestamp and SITECODE_WL.
}
\description{
This function pulls in the water level data from the NETN RAM database,
joins the location, well visit, and raw logger pressure data to calculate water level
relative to wetland surface. Function only includes data between the spring and fall
well visit per year. At least 1 site of water level data is required.
\strong{Must have the NETN RAM backend database named as a DSN.}

Note that this function has been updated to work with data collected by MX2001 water level loggers
with built-in atmospheric pressure correction, and requires the differential pressure metric.
Use package version 0.1.5 to compile water level for HOBO U20 Loggers with external
atmospheric pressure correction.

Function is primarily for internal use, and only works with data collected from 10/28/2024 and later.
Data collected prior to 10/28/2024 should be analyzed using version 0.1.5 of this package.
}
\examples{
\dontrun{
# Export growing season only data to a table
dir = c('./Water_level_data/growing_season_2025')
compile_WL_data(path = dir, year = 2025, export = TRUE, growing_season = TRUE)

# Assign output from all 2018 data, including flagged records, to global environment
wl_2025 <- compile_WL_data(year = 2025, rejected = TRUE, growing_season = FALSE,
                                   export = FALSE)

# Run for 2025 growing season data without printing messages in the console, and save output to file
wl_2025 <- compile_WL_data(year = 2025, growing_season = TRUE, export = TRUE,
                                   quietly = TRUE)
}

}
