% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NADP_precip.R
\name{get_NADP_precip}
\alias{get_NADP_precip}
\title{get_NADP_precip: Downloads and cleans hourly precipitation data from NADP}
\usage{
get_NADP_precip(
  start_date = paste0("04/01/", as.numeric(format(Sys.Date(), "\%Y"))),
  end_date = paste0("10/31/", as.numeric(format(Sys.Date(), "\%Y"))),
  stationID = "ME98",
  quietly = FALSE
)
}
\arguments{
\item{start_date}{Quoted date for first day of period (eg "04/01/2018")}

\item{end_date}{Quoted date for last day of period (eg "10/31/2018")}

\item{stationID}{Quoted name of station to pull data from (eg "ME98" is ACAD's weather station)}

\item{quietly}{\code{TRUE} or \code{FALSE}. If \code{FALSE}, code will not print progress into console.
Defaults to \code{FALSE}}
}
\value{
Returns a data frame with Date, hour (0-23), timestamp (Y-M-D H:M:S), and precip_cm
}
\description{
This function uses start date, end date and stationID to download NADP
hourly precipitation data, and converts preciptation from inches to cm. Note that
some stations may have different download links, and may not always work. The link
used to download the NADP data was generated for station: ME98.
}
\examples{
\dontrun{
precip_2018 <- get_NADP_precip(start_date='04/01/2018', end_date='10/31/2018', stationID='ME98')
}
}
