% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importWaterLevel.R
\name{importWaterLevel}
\alias{importWaterLevel}
\title{importWaterLevel: Imports and compiles views for wetland water level data package}
\usage{
importWaterLevel(
  type = "DSN",
  odbc = "RAM_BE",
  db_path = NA,
  wl_path = NA,
  new_env = TRUE,
  export_data = FALSE,
  export_path = NA,
  zip = FALSE
)
}
\arguments{
\item{type}{Select whether to use the default Data Source Named database (DSN) to import data or a
different database.
If "DSN" is selected, must specify name in odbc argument.
\describe{
\item{"DSN"}{Default. DSN database. If odbc argument is not specified, will default to "RAM_BE"}
\item{"file"}{A different database than default DSN}
}}

\item{odbc}{DSN of the database when using type = DSN. If not specified will default to "RAM_BE", which
is the back end of the MS Access RAM database.}

\item{db_path}{Quoted path of database back end file, including the name of the backend. Only needed
if type = 'file'.}

\item{wl_path}{Path and file name of csv containing corrected water level and precipitation data. 
These are generated using a script each year. Data and script are stored on the NETN Z:/ drive
under ".../Freshwater_Wetlands/ACAD_FW_Wetlands/5_Data/Wetland_Wells/Data_Analysis/"}

\item{new_env}{Logical. Specifies which environment to store views in. If \code{TRUE}(Default), stores
views in VIEWS_WL environment. If \code{FALSE}, stores views in global environment}

\item{export_data}{Logical. If TRUE, writes views to disk. If FALSE (Default), views are only
stored in specified R environment.}

\item{export_path}{Quoted path to export views to. If blank, exports to working directory.}

\item{zip}{Logical. If TRUE, exports a zip file. If FALSE (Default), exports individual csvs.}
}
\value{
Assigns RAM views to specified environment
}
\description{
This function imports water level-related tables in the wetland RAM backend and
combines them into flattened views for the data package. Each view is added to a VIEWS_WL
environment in your workspace, or to your global environment based on whether new_env = TRUE or FALSE.
}
\examples{
\dontrun{
# Import tables from database in specific folder:
importWaterLevel(type = 'file', db_path = '../data/NETN_RAM_20260731.accdb',
  wl_path = "../data/well_prec_data_2013-2025.csv")

# Import well and well visit data from default RAM_BE database
importWaterLevel(wl_path = "../data/well_prec_data_2013-2025.csv",
                 export_data = T, export_path = "../data/wetland_data_package/")
                 
#' # Import well and well visit data from default RAM_BE database, and export views as zip 
importWaterLevel(wl_path = "../data/well_prec_data_2013-2025.csv",
                 export_data = T, export_path = "../data/wetland_data_package/", zip = T)
}

}
