% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_well_data.R
\name{join_well_data}
\alias{join_well_data}
\title{join_well_data: Joins sentinal well data using data files that are exported from HoboWare
as .csv files (for internal use).}
\usage{
join_well_data(path, output = c("short", "verbose"))
}
\arguments{
\item{path}{Quoted path of the folder where the exported Hobo tables are located.}
}
\value{
Returns a wide data frame with timestamp, SITENAME_AbsPres, SITENAME_C.
}
\description{
This function finds the well data files based on a specified path
and partial matching of files based on site name, reads these files into R,
adds Well_ID and site name to each file, then joins the data to create a wide
format of the data with the timestamp as the joining column. Function requires
all 8 sites and 2 barometric loggers to run. Mostly for internal use.
}
\examples{
path='D:/NETN/Monitoring_Projects/Freshwater_Wetland/Hobo_Data/Fall_2018'
well_data<-join_well_data(path)

}
