% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NWCA16_import.R
\name{NWCA16_import}
\alias{NWCA16_import}
\title{NWCA16_import: Import NWCA 2016 data from file}
\usage{
NWCA16_import(path = NA, new_env = TRUE, zip_name = NA, ACAD_only = TRUE)
}
\arguments{
\item{path}{Quoted path of folder containing data files.}

\item{new_env}{Logical. Specifies which environment to store data frames in. If \code{TRUE}(Default), stores
data frames in NWCA16 environment. If \code{FALSE}, stores data frames in global environment}

\item{zip_name}{Quoted string ending in .zip. If specified, function looks for the specified file name and
imports .csvs from the zip file. If not specified, function looks for and imports individual csvs. Note that
this takes slightly longer than loading individual .csvs, due to the unzipping process.}

\item{ACAD_only}{Logical. If TRUE, only imports ACAD sentinel sites. If FALSE, imports full dataset.}
}
\value{
NWCA16 data files as data frames in specified environment
}
\description{
This function imports all data files from the NWCA2016 from a zip or individual files as data frames. Metadata files are not imported. Each data frame is either added to the global environment or to an environment named NWCA16, based on whether new_env = TRUE or FALSE.
}
\examples{
\dontrun{
# Import individual csvs into global environment
NWCA16_import(path = "./data/", new_env = FALSE)

# Import zipped csvs into NWCA16 environment
NWCA16_import(path = "./data/", zip_name = "NWCA2016_data.zip")
}

}
