% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NWCA16_vegMMI.R
\name{NWCA16_vegMMI}
\alias{NWCA16_vegMMI}
\title{NWCA16_vegMMI: calculate Vegetation Multi-metric Index}
\usage{
NWCA16_vegMMI()
}
\value{
NWCA16 data files as data frames in specified environment
}
\description{
This function calculates the 4 metrics used to calculate the vegetation multi-metric index (MMI) developed by Miller et al. (2016). Note that to calculate the MMI, the species data must be joined with the Coefficient of Conservatism ranking. This is only implemented for EcoRegion 82 and downloads the CoCs from the Maine Natural Areas Program's website: https://www.maine.gov/dacf/mnap/features/coc.htm. This function is only implemented for ACAD sites and will break if the url hosting the CoC rankings changes.
}
\examples{
\dontrun{

#+++++ ADD EXAMPLES +++++
# Import ACAD only data
NWCA16_import(path = "../data/NWCA16", zip_name = "NWCA2016_data.zip")

}

}
