% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NWCA21_vegMMI.R
\name{NWCA21_vegMMI}
\alias{NWCA21_vegMMI}
\title{NWCA16_vegMMI: calculate Vegetation Multi-metric Index from local JSON}
\usage{
NWCA21_vegMMI(path = NA, new_env = TRUE, export_spp = TRUE)
}
\arguments{
\item{path}{Quoted path of folder containing data files.}

\item{new_env}{Logical. Specifies which environment to store data frames in. If \code{TRUE}(Default), stores
data frames in NWCA21 environment. If \code{FALSE}, stores data frames in global environment}

\item{export_spp}{Logical. If TRUE, adds a dataframe called veg_spp_cov that contains average percent cover of all species recorded at a site to the global environment. If FALSE (default), only returns the final VMMI dataframe.}
}
\value{
Data frame with vegetation MMI metrics and ratings
}
\description{
This function calculates the 4 metrics used to calculate the vegetation multi-metric index (MMI) developed by Miller et al. (2016). Note that to calculate the MMI, the species data must be joined with the Coefficient of Conservatism ranking. This is only implemented for EcoRegion 82 and downloads the CoCs from the Maine Natural Areas Program's website: https://www.maine.gov/dacf/mnap/features/coc.htm. This function is only implemented for ACAD sites and will break if the url hosting the CoC rankings changes. This function imports the local copies of the JSON files exported from EPA's app during the NWCA2021 that are needed to calculate the vegetation MMI. Each data frame is either added to the global environment or to an environment named NWCA21, based on whether new_env = TRUE or FALSE. Note that these are preliminary data sets that have not been fully QCed by EPA, and only includes sites from ACAD. Once NWCA21 data are made public, this function will be updated to include final data sets.
}
\examples{
\dontrun{

# Calculate vegMMI using local copies of veg forms in JSON format
NWCA21_vegMMI(path = "../data/NWCA21", new_env = FALSE)

}

}
