% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hydro_site_year.R
\name{plot_hydro_site_year}
\alias{plot_hydro_site_year}
\title{plot_hydro_site_year: Plots a hydrograph by each year within a site}
\usage{
plot_hydro_site_year(df, yvar, site)
}
\arguments{
\item{df}{the name of the data frame}

\item{yvar}{Quote name of the water level field}

\item{site}{Quoted name of the site for the title}
}
\value{
Returns a panel of hydrographs during the growing season for each year
in the data frame.
}
\description{
This function creates hydrographs of each year within a site. This function
requires a data frame with the following fields: doy_h, year, and water level field.
The doy_h field is the day of year with a decimal point for hour (eg 173.09 is June 22 at 9am).
Year is a 4 digit number which is a group variable and used for facet wrap.
}
\examples{
welld<-read.csv('./Analysis/FINAL_DATA/well_prec_data_2013-2018.csv')
plot_hydro_site_year(df=welld,yvar='BIGH_WL',site='Big Heath')
plot_hydro_site_year(df=welld,yvar='DUCK_WL',site='Duck Pond')

}
