% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_well_data.R
\name{prep_well_data}
\alias{prep_well_data}
\title{prep_well_data: Prepares sentinel well data for conversion to water level using data tables stored in the NETN RAM database
(for internal use).}
\usage{
prep_well_data(
  path = NULL,
  from = 2013,
  to = 2019,
  rejected = FALSE,
  growing_season = TRUE,
  export = TRUE,
  quietly = FALSE
)
}
\arguments{
\item{path}{Quoted path of the folder where the exported Hobo tables are located.}

\item{from}{First year to start joining the data}

\item{to}{Last year to end the join}

\item{rejected}{\code{TRUE} or \code{FALSE}. If \code{TRUE} will include flagged rejected records.
If \code{FALSE} will replace the rejected flagged records with NA. Defaults to \code{FALSE}.}

\item{growing_season}{\code{TRUE} or \code{FALSE}. If \code{TRUE} will include only growing season
measurements, defined as May 15 to Oct. 1. If \code{FALSE}, will return all measurements.
If growing_season = \code{TRUE} and export = \code{TRUE}, the output filename will include "GS"
Defaults to \code{TRUE}.}

\item{export}{\code{TRUE} or \code{FALSE}. Export csv file to specified path. Defaults to \code{TRUE}.}

\item{quietly}{\code{TRUE} or \code{FALSE}. If \code{FALSE}, code will not print progress into console.
Defaults to \code{FALSE}}
}
\value{
Returns a wide data frame with timestamp, SITENAME_AbsPres, SITENAME_C.
}
\description{
This function pulls in the water level data from the NETN RAM database,
joins the location, well visit and water level data together into a wide
format of the data with the timestamp as the joining column. Function requires
all 8 sites and 2 barometric loggers to run, and only includes data between the
spring and fall well visit per year. Must have a the NETN RAM backend database named
as a DSN.Function is primarily for internal use.
}
\examples{
# Export growing season only data to a table
dir='D:/NETN/Monitoring_Projects/Freshwater_Wetland/Hobo_Data/Fall_2019'
prep_well_data(path = dir, from = 2019, to = 2019, export = TRUE, growing_season = TRUE)

# Assign output from all 2018 data to global environment
welld_2018 <- prep_well_data(from = 2018, to = 2018, export = FALSE, growing_season = FALSE)

# Run without printing messages in the console
welld_2018 <- prep_well_data(from = 2018, to = 2018, export = FALSE, quietly = TRUE)

}
