% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumSpeciesList.R
\name{sumSpeciesList}
\alias{sumSpeciesList}
\title{sumSpeciesList: compile species list for each site}
\usage{
sumSpeciesList(
  site = "all",
  panel = 1:4,
  years = 2012:format(Sys.Date(), "\%Y"),
  QAQC = FALSE,
  species_type = "all",
  include_protected = T
)
}
\arguments{
\item{site}{Character. Filter on site code. Options are "all" (default) or a vector of site codes ranging from "R-01" to "R-40".}

\item{panel}{Numeric. Filter on panel number. By default, all panels are returned, and can be filtered by numbers 1 to 4.}

\item{years}{Numeric. Filter on sample year, ranging from 2012 to 2024. By default, all years are returned. Note that years 2011, 2016, 2021, and 5-year intervals of years thereafter are EPA NWCA sites that are not}

\item{QAQC}{Logical. Include QAQC visits (TRUE) or drop QAQC visits (FALSE; default).}

\item{species_type}{Character. Options are c("all", "native", "exotic")}

\item{include_protected}{Logical. If TRUE, returns protected species. If FALSE (default), removes protected species from returned data frame.}
}
\value{
Returns a data frame with species lists for each site.
}
\description{
This function compiles species lists filtered on site, year, panel, QAQC status, and native vs exotic species.
}
\examples{
\dontrun{
# import RAM data with protected records- Note that to include protected records
# protected species records must be included
importRAM(export_protected = T)

# Compile species list for all sites and all species
spp_all <- sumSpeciesList(include_protected = T)

# Compile species list for all sites and only non-protected species
spp_public <- sumSpeciesList(include_protected = T)

# Compile for 2024 and exotic species only
spp_exo_24 <- sumSpeciesList(years = 2024, species_type = "exotic")

}

}
