% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumVegMMI.R
\name{sumVegMMI}
\alias{sumVegMMI}
\title{sumVegMMI: summarize Vegetation Multimetric Index}
\usage{
sumVegMMI(
  site = "all",
  panel = 1:4,
  years = 2012:format(Sys.Date(), "\%Y"),
  QAQC = FALSE
)
}
\arguments{
\item{site}{Character. Filter on site code. Options are "all" (default) or a vector of site codes ranging from "R-01" to "R-40".}

\item{panel}{Numeric. Filter on panel number. By default, all panels are returned, and can be filtered by numbers 1 to 4.}

\item{years}{Numeric. Filter on sample year, ranging from 2012 to 2024. By default, all years are returned. Note that years 2011, 2016, 2021, and 5-year intervals of years thereafter are EPA NWCA sites that are not}

\item{QAQC}{Logical. Include QAQC visits (TRUE) or drop QAQC visits (FALSE; default).}
}
\value{
Returns a data frame with vegetation MMIs for each site.
}
\description{
This function calculates the Vegetation Multimetric Index (VMMI) for each Rapid Assessment Monitoring (RAM) site in Acadia National Park following Miller, K.M., B.R. Mitchell, and B.J. McGill. 2016. Constructing multimetric indices and testing ability of landscape metrics to assess condition of freshwater wetlands in the Northeastern US. Ecological Indicators. 66:143-152. VMMI ratings of good, fair and poor are also included. Function can filter on site, year, and QAQC status.
}
\examples{
\dontrun{
# import RAM data with protected records- Note that to calculate VMMI correctly,
# protected species records must be included
importRAM(export_protected = T)

# Calculate veg MMI for all sites and all years
vegmmi <- sumVegMMI()

# Calculate veg MMI for 2024 only
vegmmi24 <- sumVegMMI(years = 2024)

}
}
