% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_HOBO_data.R
\name{bind_HOBO_data}
\alias{bind_HOBO_data}
\title{bind_HOBO_data: Binds sentinel well data using data files that are exported from
HoboWare as .csv files (for internal use).}
\usage{
bind_HOBO_data(path, export = TRUE)
}
\arguments{
\item{path}{Quoted path of the folder where the exported Hobo tables are located.}

\item{export}{\code{TRUE} or \code{FALSE}. Export csv file to specified path. Defaults to \code{TRUE}.}
}
\value{
Returns a long data frame with ID (blank for Access to assign autonumber), Well_ID,
timestamp, Absolute_Pressure_kPa, Degrees_C, and site.
}
\description{
This function finds the well data files based on a specified path
and partial matching of files based on site name, reads these files into R,
adds Well_ID and site name to each file, then binds them together and prepares
the data to be imported into the MS Access RAM backend table: tbl_Water_Level.
If appending to the Access table is not needed, then drop the first column 'ID',
which is left blank so that Access can assign an auto number for the primary key.
This function is mostly for internal use to generate a long format of the data
to append to the MS Access table. Function requires all 8 sites and 2 barometric
loggers to run. Mostly for internal use.
}
\examples{
\dontrun{
dir = c('C:/Water_level_data/growing_season_2019')
well_data <- bind_HOBO_data(path)
}

}
