% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_sent_WL.R
\name{compile_sent_WL}
\alias{compile_sent_WL}
\title{compile_sent_WL: Calculates water level from raw sentinal well data.}
\usage{
compile_sent_WL(df, conv_table, between_visits = TRUE)
}
\arguments{
\item{df}{data frame output from join_well_data}

\item{conv_table}{data frame listing site and well info for WL conversion}

\item{between_visits}{\code{TRUE} or \code{FALSE}. If \code{TRUE} will include only measurements
between the spring and fall well visits. If \code{FALSE}, will return all measurements included
in the data frame generated from prep_well_data(). Defaults to \code{TRUE}.}
}
\value{
Returns a data frame with the water level relative to the ground surface for each
sentinal water level logger.
}
\description{
This function uses (and requires) the output from prep_well_data() to
calculate the water level relative to the wetland surface for each sentinel site.
This should only be used for one year at a time, to ensure that the function is
using the correct well measurements to calculate water level. This function
also requires a data frame called conv_table with the following fields:
  site, wellabs, ground, corfac, baro
}
\examples{

# Compile tables
dir = c('C:/Water_level_data/growing_season_2019')
well19 <- prep_well_data(path = dir, year = 2019, growing_season = TRUE, export = FALSE)
conv_tbl_19 <- prep_conv_table(path = dir, year = 2019, visits = "fall", export = FALSE)

# Create data frame with water level relative to surface, cut at spring well visit
well19WL<-compile_sent_WL(well19, conv_tbl_19, between_visits = TRUE)

}
