% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_kpa_cm.R
\name{conv_kpa_cm}
\alias{conv_kpa_cm}
\title{conv_kpa_cm: converts absolute pressure to cm of water above the logger}
\usage{
conv_kpa_cm(df, site_code, baro)
}
\arguments{
\item{df}{data frame with pressure data}

\item{site_code}{Quoted code for site name that corresponds to a field in the data frame}

\item{baro}{Quoted field name of the atmostpheric barometric pressure correction field}
}
\value{
Returns a data frame with a new column labeled with SITECODE_cm, which represents
cm of water above the logger.
}
\description{
This function converts absolute pressure in kPa to cm of water. For this
function to work, you must have your pressure field named: SITECODE_AbsPres. The SITECODE
then needs to be defined in the function arguments.
}
\examples{
df <- data.frame(BIGH_AbsPres = c(100, 99, 98), BARO_AbsPres = c(99, 96, 94))
df2 <- conv_kpa_cm(df, site_code = 'BIGH', baro = 'BARO_AbsPres')

}
