% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_kpa_cm.R
\name{conv_kpa_cm}
\alias{conv_kpa_cm}
\title{conv_kpa_cm: converts absolute pressure to cm of water above the logger}
\usage{
conv_kpa_cm(df, pres, baro)
}
\arguments{
\item{df}{data frame with pressure data}

\item{pres}{Quoted field name that corresponds absolute pressure from a water level logger}

\item{baro}{Quoted field name of the atmostpheric barometric pressure correction field}
}
\value{
Returns a data frame with a new column labeled with SITECODE_cm, which represents
cm of water above the logger.
}
\description{
This function converts absolute pressure in kPa to cm of water.
}
\examples{
df <- data.frame(BIGH_AbsPres = c(100, 99, 98), BARO_AbsPres = c(99, 96, 94))
df2 <- conv_kpa_cm(df, pres = 'BIGH_AbsPres', baro = 'BARO_AbsPres')

}
