% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_WL.R
\name{conv_WL}
\alias{conv_WL}
\title{conv_WL: converts cm of water above the logger to water level relative to surface}
\usage{
conv_WL(df, site_code, ground, corfac)
}
\arguments{
\item{df}{data frame with pressure data}

\item{site_code}{Quoted code for site name that corresponds to a field in the data frame}

\item{ground}{numeric distance between the logger and the surface of the wetland
(Logger Length + Measuring Point to Bold - Stick Up Height).}

\item{corfac}{numeric correction factor (in cm) to correct for difference in atmospheric pressure
due to differing elevations between where the water level loggers are and where the
atmospheric pressure loggers are.}
}
\value{
Returns a data frame with a new column labeled with SITECODE_WL, which is the water level
in cm relative to the surface of the wetland. Negative values are below ground, positive are above ground.
}
\description{
This function converts cm of water above the logger to water level relative
to the surface of the weltand. For this function to work, you must have a field named: SITECODE_cm.
To get this field, should either run the conv_kpa_cm() function first, or calculate cm of water
for that site by hand. The SITECODE also needs to be defined in the function arguments.
}
\examples{
df <- data.frame(BIGH_AbsPres = c(110.1, 111.3, 110.4), BARO_AbsPres = c(102.1, 103.1, 102.8))
df <- conv_kpa_cm(df, site_code = 'BIGH', baro = 'BARO_AbsPres')
df <- conv_WL(df, site_code = 'BIGH', ground = 88.9, cor = 5.004)

}
