% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importQueries.R
\name{importQueries}
\alias{importQueries}
\title{importQueries: Imports queries from RAM front end}
\usage{
importQueries(type = c("DSN", "file"), odbc = "RAM_FE", path = NA)
}
\arguments{
\item{type}{Select whether to use the default DSN to import data or a different database
\describe{
\item{"DSN"}{Default. DSN database. If not specified, will use "NETNFVM" .}
\item{"file"}{A different database than default DSN}
}}

\item{path}{Quoted path of database backend file, including the name of the backend.}
}
\value{
Assigns database tables to global environment
}
\description{
This function imports the queries from a named ODBC datasource from
NETN forest backend databases and prepares the data for plotting in wetlandViz.
}
\examples{
# Import database in specific folder:
importQueries(type='file', path='./Data/NETN_RAM_Frontend.mdb')

# Import ODBC named database
importQueries(type='DSN', odbc="RAM_FE")

}
