% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NWCA16_vegMMI.R
\name{NWCA16_vegMMI}
\alias{NWCA16_vegMMI}
\title{NWCA16_vegMMI: calculate Vegetation Multi-metric Index}
\usage{
NWCA16_vegMMI(export_spp = FALSE)
}
\arguments{
\item{export_spp}{Logical. If TRUE, adds a dataframe called veg_spp_cov16 that contains average percent cover of all species recorded at a site to the global environment. If FALSE (default), only returns the final VMMI dataframe.}
}
\value{
Data frame with vegetation MMI metrics and ratings
}
\description{
This function calculates the 4 metrics used to calculate the vegetation multi-metric index (MMI) developed by Miller et al. (2016). Note that to calculate the MMI, the species data must be joined with the Coefficient of Conservatism ranking. This is only implemented for EcoRegion 82 and downloads the CoCs from the Maine Natural Areas Program's website: https://www.maine.gov/dacf/mnap/features/coc.htm. This function is only implemented for ACAD sites and will break if the url hosting the CoC rankings changes.
}
\examples{
\dontrun{

# Import ACAD only data and export
NWCA16_import(path = "../data/NWCA16", zip_name = "NWCA2016_data.zip")
vmmi <- NWCA16_vegMMI(export_spp = TRUE)

}

}
